moved union search function to appropriate file
